backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-seed==0.3.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
Faker==18.3.2
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
toposort==1.10
